+++ threedi_modelchecker/checks/other.py
@@ -2,6 +2,7 @@ from ..threedi_model import constants
 from ..threedi_model import models
 from .base import BaseCheck
 from .base import CheckLevel
+from .geo_query import distance
 from geoalchemy2 import functions as geo_func
 from sqlalchemy import func
 from sqlalchemy import text
@@ -49,17 +50,12 @@ class CrossSectionLocationCheck(BaseCheck):
         self.max_distance = max_distance
 
     def get_invalid(self, session):
-        epsg_code = Query(models.GlobalSetting.epsg_code).limit(1).label("epsg_code")
+        # get all channels with more than 1 cross section location
         return (
             self.to_check(session)
             .join(models.Channel)
             .filter(
-                geo_func.ST_Distance(
-                    geo_func.ST_Transform(
-                        models.CrossSectionLocation.the_geom, epsg_code
-                    ),
-                    geo_func.ST_Transform(models.Channel.the_geom, epsg_code),
-                )
+                distance(models.CrossSectionLocation.the_geom, models.Channel.the_geom)
                 > self.max_distance
             )
             .all()
@@ -150,17 +146,12 @@ class ConnectionNodesLength(BaseCheck):
     def get_invalid(self, session):
         start_node = aliased(models.ConnectionNode)
         end_node = aliased(models.ConnectionNode)
-        epsg_code = Query(models.GlobalSetting.epsg_code).limit(1).label("epsg_code")
         q = (
             Query(self.column.class_)
             .join(start_node, self.start_node)
             .join(end_node, self.end_node)
             .filter(
-                geo_func.ST_Distance(
-                    geo_func.ST_Transform(start_node.the_geom, epsg_code),
-                    geo_func.ST_Transform(end_node.the_geom, epsg_code),
-                )
-                < self.min_distance
+                distance(start_node.the_geom, end_node.the_geom) < self.min_distance
             )
         )
         return list(q.with_session(session).all())
@@ -302,6 +293,44 @@ class OpenChannelsWithNestedNewton(BaseCheck):
             f"NumericalSettings.use_of_nested_newton is switched off. "
             f"This gives convergence issues. We recommend setting use_of_nested_newton = 1."
         )
+
+
+class LinestringLocationCheck(BaseCheck):
+    """Check that linestring geometry starts / ends are close to their connection nodes
+
+    This allows for reversing the geometries. threedi-gridbuilder will reverse the geometries if
+    that lowers the distance to the connection nodes.
+    """
+
+    def __init__(self, *args, **kwargs):
+        self.max_distance = kwargs.pop("max_distance")
+        super().__init__(*args, **kwargs)
+
+    def get_invalid(self, session: Session) -> List[NamedTuple]:
+        start_node = aliased(models.ConnectionNode)
+        end_node = aliased(models.ConnectionNode)
+
+        tol = self.max_distance
+        start_point = geo_func.ST_PointN(self.column, 1)
+        end_point = geo_func.ST_PointN(self.column, geo_func.ST_NumPoints(self.column))
+
+        start_ok = distance(start_point, start_node.the_geom) <= tol
+        end_ok = distance(end_point, end_node.the_geom) <= tol
+        start_ok_if_reversed = distance(end_point, start_node.the_geom) <= tol
+        end_ok_if_reversed = distance(start_point, end_node.the_geom) <= tol
+        return (
+            self.to_check(session)
+            .join(start_node, start_node.id == self.table.c.connection_node_start_id)
+            .join(end_node, end_node.id == self.table.c.connection_node_end_id)
+            .filter(
+                ~(start_ok & end_ok),
+                ~(start_ok_if_reversed & end_ok_if_reversed),
+            )
+            .all()
+        )
+
+    def description(self) -> str:
+        return f"{self.column} does not start or end at its connection node (tolerance = {self.max_distance})"
 
 
 class BoundaryCondition1DObjectNumberCheck(BaseCheck):